refactor(channel): deduplicate zhipu credential parsing and unify plan quota cache
- Extract parseZhipuTeamCredentials to @proma/shared, removing duplicate
  implementations from channel-manager.ts and ChannelForm.tsx
- Unify plan quota cache into channel-plan-quota.ts shared module with
  in-flight request deduplication, used by both ContextUsageBadge and
  ChannelPlanQuotaBadge
- Fix ChannelPlanQuotaBadge useEffect dep to use stable keys (channel.id)
  instead of the channel object reference
- Fix indentation in testZhipuCodingTeam
- Guard inferProviderFromBaseUrl with try/catch in getChannelPlanQuota
